DataShieldMDWebApplication/DataShieldMD/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2.16
django-debug-toolbar==4.4.6
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
djoser==2.2.3
idna==3.10
jwt==1.3.1
oauthlib==3.2.2
pycparser==2.22
pyjnius==1.6.1
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
